+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### Bugs fixed
+
+* [#399](https://github.com/clojure-emacs/clojure-mode/issues/399) Fix fontification of prefix characters inside keywords.
+
 ## 5.5.1 (2016-07-25)
 
 ### Bugs fixed